+++ src/main/java/org/gridsuite/actions/server/CassandraConfig.java
@@ -6,7 +6,7 @@
  */
 package org.gridsuite.actions.server;
 
-import org.gridsuite.actions.server.repositories.ScriptContingencyListRepository;
+import org.gridsuite.actions.server.contingency.ScriptContingencyListRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
