+++ src/main/java/org/gridsuite/actions/server/contingency/ContingencyList.java
@@ -4,15 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.actions.server;
+package org.gridsuite.actions.server.contingency;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-final class ActionsApi {
+public interface ContingencyList {
 
-    private ActionsApi() {
-    }
-
-    static final String API_VERSION = "v1";
+    String getName();
 }
